+++ src/data-load.ts
@@ -9,6 +9,7 @@ import {
 } from "./interfaces.js";
 import { UserNames } from "./user-names.js";
 import { UserAvatars } from "./user-avatars.js";
+import { UserStatuses } from "./user-status.js";
 import {
   CHANNELS_DATA_PATH,
   EMOJIS_DATA_PATH,
@@ -17,6 +18,7 @@ import {
   USERS_DATA_PATH,
   USER_NAMES_DATA_PATH,
   USER_AVATARS_DATA_PATH,
+  USER_STATUS_DATA_PATH,
 } from "./config.js";
 import { retry } from "./retry.js";
 import { readJsonArraySync, readSearchDataSync } from "./big-json.js";
@@ -70,6 +72,10 @@ export async function getUsers(): Promise<Users> {
 
 export async function getEmoji(): Promise<Emojis> {
   return getFile<Emojis>(EMOJIS_DATA_PATH, {});
+}
+
+export async function getUserStatuses(): Promise<UserStatuses> {
+  return getFile<UserStatuses>(USER_STATUS_DATA_PATH, {});
 }
 
 export async function getUserAvatars(): Promise<UserAvatars> {